+++ packages/soda-hooks/src/eslint.ts
@@ -1,0 +1,16 @@
+export const additionalEffectHooks = "^useStableEffect$"
+
+export const reactHooksSettings = {
+    "react-hooks": {
+        additionalEffectHooks,
+    },
+} as const
+
+const config = [
+    {
+        name: "soda-hooks/react-hooks",
+        settings: reactHooksSettings,
+    },
+] as const
+
+export default config