+++ packages/soda-next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-next
 
+## 0.7.35
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.18.1
+
 ## 0.7.34
 
 ### Patch Changes